+++ mep3_planning/include/mep3_planning/navigate_to_pose.hpp
@@ -35,7 +35,7 @@ namespace mep3_planning
         static BT::PortsList providedPorts()
         {
             return {
-                BT::InputPort<geometry_msgs::msg::PoseStamped>("goal")};
+                BT::InputPort<BT::Pose2D>("goal")};
         }
     };
