+++ mep3_planning/src/mep3_planning.cpp
@@ -10,32 +10,6 @@
 
 // mep3_planning::BehaviorTreeRosNode::BehaviorTreeRosNode() : rclcpp::Node("behavior_tree_ros_node") {}
 
-namespace BT
-{
-    // Reference: https://www.behaviortree.dev/tutorial_03_generic_ports/
-    template <>
-    inline Pose2D convertFromString(StringView str)
-    {
-        // The next line should be removed...
-        printf("Converting string: \"%s\"\n", str.data());
-
-        // We expect real numbers separated by semicolons
-        auto parts = splitString(str, ';');
-        if (parts.size() != 3)
-        {
-            throw RuntimeError("invalid input)");
-        }
-        else
-        {
-            Pose2D output;
-            output.x = convertFromString<double>(parts[0]);
-            output.y = convertFromString<double>(parts[1]);
-            output.theta = convertFromString<double>(parts[2]);
-            return output;
-        }
-    }
-}
-
 int main(int argc, char **argv)
 {
     // Load strategy from file
@@ -51,20 +25,16 @@ int main(int argc, char **argv)
         return 1;
     }
 
-    auto node = rclcpp::Node::make_shared("navigate_to_pose");
+    auto node = rclcpp::Node::make_shared("mep3_planning");
 
     BT::BehaviorTreeFactory factory;
-    // BT::SharedLibrary loader;
-
-    // factory.registerFromPlugin(loader.getOSName("mr2_battery_checker_bt_node"));
-    // factory.registerFromPlugin(loader.getOSName("mr2_patrol_bt_node"));
-    // factory.registerFromPlugin(loader.getOSName("mr2_recharge_bt_node"));
-    // factory.registerFromPlugin(loader.getOSName("mr2_move_bt_node"));
-    // factory.registerFromPlugin(loader.getOSName("mr2_get_waypoint_bt_node"));
 
     auto blackboard = BT::Blackboard::create();
     blackboard->set("node", node);
     BT::Tree tree = factory.createTreeFromFile(tree_file, blackboard);
+    BT::StdCoutLogger logger_cout(tree);
+
+    rclcpp::init(argc, argv);
 
     bool finish = false;
     while (!finish && rclcpp::ok())
@@ -73,24 +43,6 @@ int main(int argc, char **argv)
 
         rclcpp::spin_some(node);
     }
-
     rclcpp::shutdown();
     return 0;
-
-    // // Create a behavior tree
-    // BT::BehaviorTreeFactory factory;
-    // factory.registerNodeType<NavigateToAction>("NavigateToAction");
-    // auto tree = factory.createTreeFromFile(tree_file);
-    // BT::StdCoutLogger logger_cout(tree);
-
-    // // Pass the ROS node to Behavior Tree nodes
-    // rclcpp::init(argc, argv);
-    // mep3_planning::BehaviorTreeRosNode::SharedPtr behavior_tree_ros_node = std::make_shared<mep3_planning::BehaviorTreeRosNode>();
-    // for (auto &node : tree.nodes)
-    //     if (mep3_planning::InitInterface *behavior_tree_node = dynamic_cast<mep3_planning::InitInterface *>(node.get()))
-    //         behavior_tree_node->init(behavior_tree_ros_node);
-    // tree.tickRoot();
-    // rclcpp::shutdown();
-
-    // return 0;
 }
