+++ mep3_planning/src/navigate_to_pose.cpp
@@ -10,6 +10,32 @@
 
 #include "behaviortree_cpp_v3/behavior_tree.h"
 
+namespace BT
+{
+    // Reference: https://www.behaviortree.dev/tutorial_03_generic_ports/
+    template <>
+    inline Pose2D convertFromString(StringView str)
+    {
+        // The next line should be removed...
+        printf("Converting string: \"%s\"\n", str.data());
+
+        // We expect real numbers separated by semicolons
+        auto parts = splitString(str, ';');
+        if (parts.size() != 3)
+        {
+            throw RuntimeError("invalid input)");
+        }
+        else
+        {
+            Pose2D output;
+            output.x = convertFromString<double>(parts[0]);
+            output.y = convertFromString<double>(parts[1]);
+            output.theta = convertFromString<double>(parts[2]);
+            return output;
+        }
+    }
+}
+
 namespace mep3_planning
 {
 
@@ -25,10 +51,13 @@ namespace mep3_planning
 
     void NavigateToPose::on_tick()
     {
-        geometry_msgs::msg::PoseStamped goal;
+        BT::Pose2D goal;
         getInput("goal", goal);
 
-        goal_.pose = goal;
+        goal_.pose.header.frame_id = "map";
+        goal_.pose.header.stamp = node_->get_clock()->now();
+        goal_.pose.pose.position.x = goal.x;
+        goal_.pose.pose.position.y = goal.y;
     }
 
     BT::NodeStatus NavigateToPose::on_success()
